fix: risolto errore 500 su /api/admin/notifications/cleanup
- AdminNotificationRepository: aggiunto @Modifying e @Transactional, cambiato ritorno da long a int
- AdminNotificationService: allineato tipo ritorno cleanupArchivedOlderThanDays() da long a int
- AdminDashboardController: allineato tipo variabile deleted da long a int
- AdminNotificationCleanupService: allineato tipo variabili deletedResolved e totalDeleted da long a int (2 occorrenze)
- AdminNotificationController: già aggiornato in precedenza (usa int e chiama direttamente repository)

Risolve errore in produzione:
'Null return value from advice does not match primitive return type for: public abstract long'

Causa: metodo delete senza @Modifying eseguiva SELECT invece di DELETE, restituendo null che non può essere assegnato a long primitivo.

Soluzione: uniformato tutto a int + @Modifying + @Transactional, conforme a Spring Data JPA e pattern già usato in AdminActionLogRepository.

+++ src/main/java/com/funkard/admin/service/AdminNotificationCleanupService.java
@@ -28,9 +28,9 @@ public class AdminNotificationCleanupService {
             LocalDateTime cutoff = LocalDateTime.now().minusDays(30);
             
             // 1. Pulisce notifiche risolte da più di 30 giorni
-            long deletedResolved = repository.deleteByArchivedTrueAndArchivedAtBefore(cutoff);
+            int deletedResolved = repository.deleteByArchivedTrueAndArchivedAtBefore(cutoff);
             
-            long totalDeleted = deletedResolved;
+            int totalDeleted = deletedResolved;
             
             if (totalDeleted > 0) {
                 System.out.println("🧹 [CLEANUP] Pulizia completata:");
@@ -52,11 +52,11 @@ public class AdminNotificationCleanupService {
         try {
             LocalDateTime cutoff = LocalDateTime.now().minusDays(daysOld);
             
-            long deletedResolved = repository.deleteByArchivedTrueAndArchivedAtBefore(cutoff);
+            int deletedResolved = repository.deleteByArchivedTrueAndArchivedAtBefore(cutoff);
             
             System.out.println("🧹 [MANUAL CLEANUP] Rimossi " + deletedResolved + " record più vecchi di " + daysOld + " giorni");
             
-            return (int) deletedResolved;
+            return deletedResolved;
             
         } catch (Exception e) {
             System.err.println("❌ [MANUAL CLEANUP] Errore: " + e.getMessage());
